+++ src/entities/post/api/postApi.ts
@@ -1,4 +1,4 @@
-import { Post } from "@/types/api";
+import { Post } from "@/entities/post/model/types";
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 
 export const postsApi = createApi({
